Renamed random.h to random_quda.h to avoid naming collisions.
