+++ lib/pgauge_init.cu
@@ -7,7 +7,7 @@
 #include <comm_quda.h>
 #include <unitarization_links.h>
 #include <pgauge_monte.h>
-#include <random.h>
+#include <random_quda.h>
 #include <cub/cub.cuh>
 #include <index_helper.cuh>
