+++ tests/gauge_alg_test.cpp
@@ -11,7 +11,7 @@
 #include <gauge_tools.h>
 
 #include <pgauge_monte.h>
-#include <random.h>
+#include <random_quda.h>
 #include <unitarization_links.h>
